feature wip stats menu in inventory
* fixes to reload the ui
* minor fixes for getting values from config
* atm showing slotid and what item is currently equipped (just armor,
  shields, accessories, clothing)

+++ src/events/keymanager.h
@@ -77,7 +77,7 @@ public:
             } else if (key == RE::BSWin32KeyboardDevice::Key::kEscape) {
                 showHandler->CloseAllWindows();
                 break;
-            } else if (key == *Settings::openFactionMenuButton) {
+            } else if (key == static_cast<uint32_t>(*Settings::openFactionMenuButton)) {
                 showHandler->HandleNextMenuButtonPress();
             }
         }
